refactoring code and something more
