+++ src/components/Dashboard/Login/Form.jsx
@@ -4,7 +4,7 @@ import { FaM } from "react-icons/fa6";
 import { AiFillLock } from "react-icons/ai";
 import { Button, TextInput } from "flowbite-react";
 import { useForm } from "react-hook-form";
-import { errorStore } from "../../../store/useStore";
+import { errorStore } from "../../../store/store";
 
 const Form = ({data}) => {
     const { register, handleSubmit } = useForm();
