+++ src/components/Dashboard/Table/Table.jsx
@@ -9,14 +9,14 @@ import {
   errorStore,
   successStore,
   tokenStatusStore,
-} from "../../../store/useStore";
+} from "../../../store/store";
 import {
   useCreate,
   useUpdate,
   useDelete,
   useGet,
   checkTokenStatus,
-} from "../../../hooks/queryData";
+} from "../../../hooks/useQueryData";
 import { useEffect, useMemo, useState } from "react";
 import {
   validateCreateData,
@@ -45,7 +45,10 @@ export const TableContainer = ({ keyword }) => {
     }
   }, [tokenState, navigate]);
 
-  const columns = useMemo(() => getColumns(validationErrors, setValidationErrors), [validationErrors, setValidationErrors]);
+  const columns = useMemo(
+    () => getColumns(validationErrors, setValidationErrors),
+    [validationErrors, setValidationErrors]
+  );
 
   //call GET hook
   const {
