+++ src/pages/index.jsx
@@ -1,0 +1,4 @@
+export { DashboardPage } from "./DashboardPage";
+export { LoginPage } from "./LoginPage";
+export { NotFoundPage } from "./NotFoundPage";
+export { ProductsPage } from "./ProductsPage";
