+++ src/store/store.js
@@ -22,6 +22,7 @@ const errorStore = createStore("error-storage", {
     statusCode: null,
     message: null,
 });
+
 const successStore = createStore("success-storage", {
     statusCode: null,
     message: null,
